refactor(core): share predicate expression helpers
