+++ packages/core/src/transformers/PredicateReachabilityAnalyzer.ts
@@ -18,7 +18,6 @@ import {
     FunctionCall,
     InlineQuery,
     JsonPredicateExpression,
-    ParenExpression,
     TupleExpression,
     TypeValue,
     UnaryExpression,
@@ -26,9 +25,19 @@ import {
     ValueList
 } from "../models/ValueComponent";
 import { SelectQueryParser } from "../parsers/SelectQueryParser";
-import { ValueParser } from "../parsers/ValueParser";
 import { formatSqlComponent, SqlComponentFormatOptions } from "./SqlComponentFormatter";
 import { SelectOutputCollector, SelectOutputColumn } from "./SelectOutputCollector";
+import {
+    cloneColumnReference,
+    cloneValueComponent,
+    collectTopLevelAndTerms,
+    columnReferenceText,
+    identifiersEqual,
+    isBinaryOperator,
+    normalizeIdentifier,
+    sameColumnReference,
+    unwrapParens
+} from "./PredicateExpressionUtils";
 
 export type PredicateReachabilityInput = string | SelectQuery | SimpleSelectQuery;
 
@@ -123,64 +132,6 @@ interface TargetColumnResolution {
     sourceColumn: ColumnReference;
     targetColumn: ColumnReference;
 }
-
-const normalizeIdentifier = (value: string): string => value.trim().toLowerCase();
-const isCaseSensitiveIdentifier = (value: string): boolean => /[A-Z]/.test(value.trim());
-const identifiersEqual = (left: string, right: string): boolean => {
-    const trimmedLeft = left.trim();
-    const trimmedRight = right.trim();
-    return isCaseSensitiveIdentifier(trimmedLeft) || isCaseSensitiveIdentifier(trimmedRight)
-        ? trimmedLeft === trimmedRight
-        : trimmedLeft.toLowerCase() === trimmedRight.toLowerCase();
-};
-
-const unwrapParens = (expression: ValueComponent): ValueComponent => {
-    let candidate = expression;
-    while (candidate instanceof ParenExpression) {
-        candidate = candidate.expression;
-    }
-    return candidate;
-};
-
-const isBinaryOperator = (expression: ValueComponent, operator: string): expression is BinaryExpression => {
-    const candidate = unwrapParens(expression);
-    return candidate instanceof BinaryExpression
-        && candidate.operator.value.trim().toLowerCase() === operator;
-};
-
-const collectTopLevelAndTerms = (expression: ValueComponent): ValueComponent[] => {
-    const candidate = unwrapParens(expression);
-    if (!isBinaryOperator(candidate, "and")) {
-        return [expression];
-    }
-
-    return [
-        ...collectTopLevelAndTerms(candidate.left),
-        ...collectTopLevelAndTerms(candidate.right)
-    ];
-};
-
-const columnReferenceText = (reference: ColumnReference): string => {
-    const namespace = reference.getNamespace();
-    return namespace ? `${namespace}.${reference.column.name}` : reference.column.name;
-};
-
-const sameColumnReference = (left: ColumnReference, right: ColumnReference): boolean => {
-    return identifiersEqual(left.column.name, right.column.name)
-        && identifiersEqual(left.getNamespace(), right.getNamespace());
-};
-
-const cloneValueComponent = (
-    expression: ValueComponent,
-    options: SqlComponentFormatOptions
-): ValueComponent => {
-    return ValueParser.parse(formatSqlComponent(expression, options));
-};
-
-const cloneColumnReference = (reference: ColumnReference): ColumnReference => {
-    const namespaces = reference.namespaces?.map(namespace => namespace.name) ?? null;
-    return new ColumnReference(namespaces, reference.column.name);
-};
 
 export class PredicateReachabilityAnalyzer {
     public analyze(
